CORE-2509. Took the opportunity to fix several data types.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@42653 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/trace/TraceObjects.cpp
@@ -215,7 +215,7 @@ const char* TraceSQLStatementImpl::getPlan()
 		char buff;
 		m_plan = &buff;
 
-		const int len = DSQL_get_plan_info(JRD_get_thread_data(),
+		const size_t len = DSQL_get_plan_info(JRD_get_thread_data(),
 			m_stmt, sizeof(buff), &m_plan);
 
 		if (len)
